requirements.txt
#Langchain core
langchain
langchain-core

#openAI integration
langchain-openai 
openai 

# Anthropic integration
langchain-anthropic 

#google gemini (PaLm) Integration
langchain-google-genai 
google-generativeai 

#Hugging Face Integration
langchain-huggingface 
transformers
huggingface-hub 

# environment variable management
python-dotenv
#Machine learning utilities
numpy
scikit-learn 

